Add Apache License 2.0 copyright notice to ZeRO (#162)
## Title

- Add ColossalAI Apache License 2.0 copyright notice to modified script 

## Description

- This PR adds the Apache License 2.0 copyright notice to the modified
script from the ColossalAI project. The original script is under the
Apache License 2.0, and as per the license requirements, we have
included the original copyright notice, as well as an additional comment
mentioning the modifications made by our organization.

##  Changes:
- Added Apache License 2.0 copyright notice to the modified script
- Included a comment to indicate the modifications made by our
organization

Please review and let me know if any changes are required. Thanks!