[zero] Suggests a minor change to confusing variable names in the ZeRO optimizer (#183)
## Title

- [zero] Suggests a minor change to confusing variable names in the ZeRO
optimizer

## Description

It seems that the variable names related to the mixed precision
parameter group do not comprehensively cover its characteristics, so I
suggest a few changes. These changes are very trivial, but hopefully
they will alleviate some of the confusion for beginners like me.

Currently, the entire parameter group is named `fp16_param_groups`, and
the parts managed by the gpu at the current rank are described as
`fp32_flat_param_groups_of_current_rank`. This state perfectly
represents the characteristics when the master weight is a half-tensor
or the dtype specified in the `__init__`method is fp16. In other cases,
however, its characteristics do not correspond to the variable it.

I would like to propose an alternative term, `working_param` and
`master_param`. The term is more closely related to the concept of the
mixed-precision training context. Using `working_param` and
`master_weight` would create a clear distinction between the two types
of parameters and help avoid confusion.

To summarize my suggestions:
- `fp16` -> `working`
- `fp32` -> `master`


## Linked Issues

- N/A

## Reference

- https://github.com/hpcaitech/ColossalAI/pull/3173